requirements.txt
beautifulsoup4==4.9.1
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1598370769933/work
chardet==3.0.4
click==7.1.2
cryptography @ file:///tmp/build/80754af9/cryptography_1598892038851/work
Flask==1.1.2
html5lib @ file:///tmp/build/80754af9/html5lib_1593446221756/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
PySocks==1.7.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
six==1.15.0
soupsieve==2.0.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
webencodings==0.5.1
Werkzeug==1.0.1
